+++ OAuth/ResourceOwner/LinkedinResourceOwner.php
@@ -11,51 +11,51 @@
 
 namespace HWI\Bundle\OAuthBundle\OAuth\ResourceOwner;
 
+use Buzz\Message\RequestInterface as HttpRequestInterface;
+
 /**
  * LinkedinResourceOwner
  *
  * @author Francisco Facioni <fran6co@gmail.com>
+ * @author Joseph Bielawski <stloyd@gmail.com>
  */
-class LinkedinResourceOwner extends GenericOAuth1ResourceOwner
+class LinkedinResourceOwner extends GenericOAuth2ResourceOwner
 {
     /**
      * {@inheritDoc}
      */
     protected $options = array(
-        'authorization_url'   => 'https://www.linkedin.com/uas/oauth/authenticate',
-        'request_token_url'   => 'https://api.linkedin.com/uas/oauth/requestToken',
-        'access_token_url'    => 'https://api.linkedin.com/uas/oauth/accessToken',
-        'infos_url'           => 'http://api.linkedin.com/v1/people/~:(id,formatted-name)',
-
-        'realm'               => 'http://api.linkedin.com',
+        'authorization_url'   => 'https://www.linkedin.com/uas/oauth2/authorization',
+        'access_token_url'    => 'https://www.linkedin.com/uas/oauth2/accessToken',
+        'infos_url'           => 'https://api.linkedin.com/v1/people/~:(id,formatted-name,email-address,picture-url)?format=json',
+        'csrf'                => true,
     );
 
     /**
      * {@inheritDoc}
      */
     protected $paths = array(
-        'identifier' => 'id',
-        'nickname'   => 'formattedName',
-        'realname'   => 'formattedName',
+        'identifier'     => 'id',
+        'nickname'       => 'formattedName',
+        'realname'       => 'formattedName',
+        'email'          => 'emailAddress',
+        'profilepicture' => 'pictureUrl',
     );
 
     /**
      * {@inheritDoc}
      */
-    protected function httpRequest($url, $content = null, $parameters = array(), $headers = array(), $method = null)
+    protected function doGetTokenRequest($url, array $parameters = array())
     {
-        $headers[] = 'x-li-format: json';
-
-        return parent::httpRequest($url, $content, $parameters, $headers, $method);
+        return $this->httpRequest($this->normalizeUrl($url, $parameters), null, array(), HttpRequestInterface::METHOD_POST);
     }
 
     /**
-     * Add scope (Required by linkedin API if email address is needed)
-     *
      * {@inheritDoc}
      */
-    protected function getRequestToken($redirectUri, array $extraParameters = array())
+    protected function doGetUserInformationRequest($url, array $parameters = array())
     {
-        return parent::getRequestToken($redirectUri, array_merge(array('scope' => $this->getOption('scope')), $extraParameters));
+        // LinkedIn uses different variable as they still support OAuth1.0a
+        return parent::doGetUserInformationRequest(str_replace('access_token', 'oauth2_access_token', $url), $parameters);
     }
 }